html/includes/properties/Dynamic_FileUpload_Property.php: 	#  support magic link in file & text upload properties html/includes/properties/Dynamic_TextUpload_Property.php: 	#  support magic link in file & text upload properties

+++ html/includes/properties/Dynamic_TextUpload_Property.php
@@ -48,14 +48,32 @@ class Dynamic_TextUpload_Property extends Dynamic_Property
         if (!empty($_FILES) && !empty($_FILES[$upname]) && !empty($_FILES[$upname]['tmp_name'])
             // is_uploaded_file() : PHP 4 >= 4.0.3
             && is_uploaded_file($_FILES[$upname]['tmp_name']) && $_FILES[$upname]['size'] > 0 && $_FILES[$upname]['size'] < 1000000) {
-            // this doesn't work on some configurations
-            //$this->value = join('', @file($_FILES[$upname]['tmp_name']));
-            $tmpdir = xarCoreGetVarDirPath();
-            $tmpdir .= '/cache/templates';
-            $tmpfile = tempnam($tmpdir, 'dd');
-            if (move_uploaded_file($_FILES[$upname]['tmp_name'], $tmpfile) && file_exists($tmpfile)) {
-                $this->value = join('', file($tmpfile));
-                unlink($tmpfile);
+
+            // if the uploads module is hooked (to be verified and set by the calling module)
+            if (xarVarGetCached('Hooks.uploads','ishooked')) {
+                $magicLink = xarModAPIFunc('uploads',
+                                           'user',
+                                           'uploadmagic',
+                                           array('uploadfile'=>$upname,
+                                                 'mod'=>'dynamicdata',
+                                                 'modid'=>0,
+                                                 'utype'=>'file'));
+                if (!empty($value)) {
+                    $value .= ' ' . $magicLink;
+                } else {
+                    $value = $magicLink;
+                }
+                $this->value = $value;
+            } else {
+                // this doesn't work on some configurations
+                //$this->value = join('', @file($_FILES[$upname]['tmp_name']));
+                $tmpdir = xarCoreGetVarDirPath();
+                $tmpdir .= '/cache/templates';
+                $tmpfile = tempnam($tmpdir, 'dd');
+                if (move_uploaded_file($_FILES[$upname]['tmp_name'], $tmpfile) && file_exists($tmpfile)) {
+                    $this->value = join('', file($tmpfile));
+                    unlink($tmpfile);
+                }
             }
         } elseif (!empty($value)) {
             $this->value = $value;
